+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.8.1] - 2024-05-27
-### Changed
-- `CatalogBuilder::new_chained()` gets the `chained_catalog` argument by value, not by reference
+### Fixed
+- Fixed pedantic linter warnings
 
 ## [0.8.0] - 2023-11-27
 ### Added