+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] - 2024-08-15
+### Fixed
+- `Catalog::builders_for_with_meta()` works corectly for chained catalogs
+
+
 ## [0.9.0] - 2024-07-29
 ### Added
 - It's now possible to associate custom static metadata with builders: